src/bilanciafrigo_server_bot/requirements.txt
python-telegram-bot==20.0
paho-mqtt
